set enemy patrols
