+++ OoO_Game/Assets/Prefabs/advanced_enemy.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 2697356933772033103}
   - component: {fileID: 7704964364191713298}
   - component: {fileID: 4637087719488975306}
+  - component: {fileID: 2000330552331693584}
   m_Layer: 0
   m_Name: advanced_enemy
   m_TagString: enemy
@@ -28,7 +29,7 @@ Transform:
   m_GameObject: {fileID: 5539990592294687455}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -3.1242201, y: 1.8056775, z: -0.09394269}
+  m_LocalPosition: {x: -3.1242201, y: 1.8056775, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -100,6 +101,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   health: 300
   speed: 1
+  velocity: {x: 0, y: 1, z: 0}
 --- !u!61 &4637087719488975306
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -146,3 +148,30 @@ BoxCollider2D:
   m_AutoTiling: 0
   m_Size: {x: 0.74, y: 0.45}
   m_EdgeRadius: 0
+--- !u!50 &2000330552331693584
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5539990592294687455}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 0
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
