--- OoO_Game/Assets/Scripts/basic_enemy.cs
@@ -1,22 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class basic_enemy : MonoBehaviour
-{
-
-    public int health = 100;
-    public int speed = 1;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
