+++ OoO_Game/Assets/Scripts/camera.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54c7450e4f97b2f4197fc626a1c1c1b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
