+++ OoO_Game/Assets/Scripts/enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a27c85cfa602e81479b957babeb6c79b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
