+++ OoO_Game/Assets/Scripts/enemies/abstract_enemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81f09f46e0049134b82c22c2e0fd433a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
