+++ OoO_Game/Assets/Scripts/enemies/basic_enemy.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class basic_enemy : abstract_enemy
+{
+    private void Start()
+    {
+        health = 100;
+        speed = 5;
+        velocity = new Vector3(0.0f, 1.0f, 0.0f);
+    }
+}
